Fixed source scope in multi-module runs #1053
