+++ src/PSRule/Definitions/Selectors/SelectorVisitor.cs
@@ -1,22 +1,28 @@
-﻿// Copyright (c) Microsoft Corporation.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using System;
 using System.Diagnostics;
 using PSRule.Definitions.Expressions;
+using PSRule.Pipeline;
 using PSRule.Resources;
 
 namespace PSRule.Definitions.Selectors
 {
+    internal interface ISelector : ILanguageBlock
+    {
+
+    }
+
     [DebuggerDisplay("Id: {Id}")]
-    internal sealed class SelectorVisitor
+    internal sealed class SelectorVisitor : ISelector
     {
         private readonly LanguageExpressionOuterFn _Fn;
 
-        public SelectorVisitor(string module, string id, LanguageIf expression)
+        public SelectorVisitor(ResourceId id, SourceFile source, LanguageIf expression)
         {
-            Module = module;
             Id = id;
+            Source = source;
             InstanceId = Guid.NewGuid();
             var builder = new LanguageExpressionBuilder();
             _Fn = builder.Build(expression);
@@ -24,13 +30,19 @@ namespace PSRule.Definitions.Selectors
 
         public Guid InstanceId { get; }
 
-        public string Module { get; }
+        public ResourceId Id { get; }
 
-        public string Id { get; }
+        public SourceFile Source { get; }
+
+        [Obsolete("Use Source property instead.")]
+        string ILanguageBlock.SourcePath => Source.Path;
+
+        [Obsolete("Use Source property instead.")]
+        string ILanguageBlock.Module => Source.Module;
 
         public bool Match(object o)
         {
-            var context = new ExpressionContext(Module);
+            var context = new ExpressionContext(Source);
             context.Debug(PSRuleResources.SelectorMatchTrace, Id);
             return _Fn(context, o).GetValueOrDefault(false);
         }
