+++ tests/PSRule.Tests/SelectorTests.cs
@@ -1502,8 +1502,8 @@ namespace PSRule
             context = new RunspaceContext(PipelineContext.New(GetOption(), null, null, PipelineHookActions.BindTargetName, PipelineHookActions.BindTargetType, PipelineHookActions.BindField, new OptionContext(), null), null);
             context.Init(source);
             context.Begin();
-            var selector = HostHelper.GetSelector(source, context).ToArray();
-            return new SelectorVisitor(null, name, selector.FirstOrDefault(s => s.Name == name).Spec.If);
+            var selector = HostHelper.GetSelector(source, context).ToArray().FirstOrDefault(s => s.Name == name);
+            return new SelectorVisitor(selector.Id, selector.Source, selector.Spec.If);
         }
 
         private static string GetSourcePath(string fileName)
