+++ tests/PSRule.Tests/TestModule9a/Rules.Rule.ps1
@@ -1,0 +1,12 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+# Synopsis: Test scoped configuration.
+Rule 'M9ATestConfig' {
+    $Configuration.ConfigA -eq 'ValueA'
+}
+
+# Synopsis: Test scoped configuration.
+Rule 'M9ATestConfig2' {
+    $Null -eq $Configuration.ConfigB
+}
