+++ tests/PSRule.Tests/TestModule9b/Config.Rule.yaml
@@ -1,0 +1,21 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+---
+apiVersion: github.com/microsoft/PSRule/v1
+kind: ModuleConfig
+metadata:
+  name: TestModule9b
+spec:
+  rule:
+    baseline: Default
+
+---
+# Synopsis: The default baseline for TestModule9b.
+apiVersion: github.com/microsoft/PSRule/v1
+kind: Baseline
+metadata:
+  name: Default
+spec:
+  configuration:
+    ConfigB: 'ValueB'
